requirements.txt
aiofiles==23.2.1
blinker==1.5
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
prettytable==3.9.0
priority==2.0.0
pyecharts==2.0.4
quart==0.18.4
quart-cors==0.7.0
quickchart-io==2.0.0
requests==2.31.0
simplejson==3.19.1
urllib3==2.0.5
wcwidth==0.2.6
Werkzeug==2.3.7
wsproto==1.2.0
